+++ Libplanet.Extensions.Cocona/Commands/KeyCommand.cs
@@ -165,6 +165,60 @@ namespace Libplanet.Extensions.Cocona.Commands
             }
         }
 
+        [Command(Description = "Sign a message.")]
+        public void Sign(
+            [Argument("KEY-ID", Description = "A key UUID to sign.")]
+            Guid keyId,
+            [Argument(
+                "FILE",
+                Description = "A path of the file to sign. If you pass '-' dash character, " +
+                              "it will receive the message to sign from stdin."
+            )]
+            string path,
+            PassphraseParameters passphrase,
+            [Option(Description = "A path of the file to save the signature. " +
+                                  "If you pass '-' dash character, it will print to stdout " +
+                                  "as raw bytes not hexadecimal string or else. " +
+                                  "If this option isn't given, it will print hexadecimal string " +
+                                  "to stdout as default behaviour.")]
+            string? binaryOutput = null
+        )
+        {
+            PrivateKey key = UnprotectKey(keyId, passphrase);
+
+            byte[] message;
+            if (path == "-")
+            {
+                // Stream for stdin does not support .Seek()
+                using MemoryStream buffer = new MemoryStream();
+                using (Stream stream = Console.OpenStandardInput())
+                {
+                    stream.CopyTo(buffer);
+                }
+
+                message = buffer.ToArray();
+            }
+            else
+            {
+                message = File.ReadAllBytes(path);
+            }
+
+            var signature = key.Sign(message);
+            if (binaryOutput is null)
+            {
+                Console.WriteLine(ByteUtil.Hex(signature));
+            }
+            else if (binaryOutput == "-")
+            {
+                using Stream stdout = Console.OpenStandardOutput();
+                stdout.Write(signature, 0, signature.Length);
+            }
+            else
+            {
+                File.WriteAllBytes(binaryOutput, signature);
+            }
+        }
+
         public PrivateKey UnprotectKey(
             Guid keyId,
             PassphraseParameters passphrase,